fix: reconfigure
